+++ input/ball/red eye ball_sr.png
