+++ input/bottom lid/tilted bottom_r.png
