+++ input/eye color/cyan big.png
