+++ input/eye color/cyan small.png
