+++ input/eye color/pink small.png
