+++ input/eye color/purple big_r.png
