+++ input/eye color/purple small.png
