+++ input/eye color/red big_sr.png
