+++ input/eye color/red small.png
