+++ input/eye color/yellow big.png
