+++ input/eye color/yellow small.png
