+++ input/shine/shapes.png
