+++ input/top lid/tilted top_r.png
