+++ output/_metadata.json
@@ -1,0 +1,1 @@
+[{"hash":"11213842516372","decodedHash":[{"1":1},{"2":1},{"3":8},{"4":2},{"5":1},{"6":3},{"7":2}],"edition":1,"date":1631025395161,"attributes":[{"id":1,"layer":"background","name":"black","rarity":"original"},{"id":1,"layer":"ball","name":"red eye ball","rarity":"super rare"},{"id":8,"layer":"eye color","name":"purple small","rarity":"original"},{"id":2,"layer":"iris","name":"medium","rarity":"original"},{"id":1,"layer":"shine","name":"shapes","rarity":"original"},{"id":3,"layer":"bottom lid","name":"tilted bottom","rarity":"rare"},{"id":2,"layer":"top lid","name":"low top","rarity":"original"}]}]
